+++ packages/terra-modal-manager/tests/nightwatch/DemoContainer.jsx
@@ -14,6 +14,8 @@ class DemoContainer extends React.Component {
     this.goBack = this.goBack.bind(this);
     this.maximize = this.maximize.bind(this);
     this.minimize = this.minimize.bind(this);
+    this.requestFocus = this.requestFocus.bind(this);
+    this.releaseFocus = this.releaseFocus.bind(this);
   }
 
   disclose() {
@@ -50,6 +52,14 @@ class DemoContainer extends React.Component {
     this.props.app.minimize();
   }
 
+  requestFocus() {
+    this.props.app.requestFocus();
+  }
+
+  releaseFocus() {
+    this.props.app.releaseFocus();
+  }
+
   render() {
     const { app, identifier } = this.props;
 
@@ -59,12 +69,17 @@ class DemoContainer extends React.Component {
         <br />
         <h4>id: {identifier}</h4>
         <br />
+        {app && app.releaseFocus ? <h4>Modal has lost focus!</h4> : null }
+        {app && app.requestFocus ? <h4>Modal has gained focus!</h4> : null }
+        <br />
         <button className="disclose" onClick={this.disclose}>Disclose</button>
         {app && app.dismiss ? <button className="dismiss" onClick={this.dismiss}>Dismiss</button> : null }
         {app && app.closeDisclosure ? <button className="close-disclosure" onClick={this.closeDisclosure}>Close Disclosure</button> : null }
         {app && app.goBack ? <button className="go-back" onClick={this.goBack}>Go Back</button> : null }
         {app && app.maximize ? <button className="maximize" onClick={this.maximize}>Maximize</button> : null }
         {app && app.minimize ? <button className="minimize" onClick={this.minimize}>Minimize</button> : null }
+        {app && app.requestFocus ? <button className="requestFocus" onClick={this.requestFocus}>Request Focus</button> : null }
+        {app && app.releaseFocus ? <button className="releaseFocus" onClick={this.releaseFocus}>Release Focus</button> : null }
       </div>
     );
   }