+++ packages/terra-modal/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added isFocused prop to support trapping and releasing focus in the modal.
 
 1.2.0 - (July 26, 2017)
 ------------------